backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
python-igraph==0.11.8
lightgbm==4.5.0
shap==0.46.0
ibm-watsonx-ai==1.1.24
fpdf2==2.8.2
python-multipart==0.0.20
python-dotenv==1.0.1
scikit-learn==1.6.1
numpy==1.26.4
